SIMD code hopefully now also works on 64bit
git-svn-id: https://pure-data.svn.sourceforge.net/svnroot/pure-data/trunk@13045 df7eb9d6-776f-4174-a42d-0ffd8d5d5b5d

+++ externals/zexy/src/0x260x260x7e.c
@@ -20,7 +20,7 @@
   1302:forum::für::umläute:2000
 */
 
-#include "zexy.h"
+#include "zexySIMD.h"
 
 /* ------------------------ logical~ ----------------------------- */
 
@@ -125,7 +125,7 @@ static t_int *scalarandand_tilde_perf8(t_int *w)
 }
 
 #ifdef __SSE__
-static long l_bitmask[]={0x7fffffff, 0x7fffffff, 0x7fffffff, 0x7fffffff};
+static int l_bitmask[]={0x7fffffff, 0x7fffffff, 0x7fffffff, 0x7fffffff};
 
 static t_int *andand_tilde_performSSE(t_int *w)
 {
